refactor(weave): move shared session test fixtures into conftest
Per review feedback (PR #6846): _reset_contextvars and otel_spans were
duplicated across test_session.py, test_session_otel.py, and
test_session_settings.py. Moved both into tests/session/conftest.py;
_reset_contextvars stays autouse so existing behavior is unchanged.
Dropped now-unused imports in each test file.
